requirements.txt
beautifulsoup4==4.12.3
openai==1.57.1
python-dotenv==1.0.1
streamlit==1.40.2